requirements.txt
apscheduler>=3.2.0
docopt>=0.6.2
flask>=0.10.1
futures>=3.0.3
humanfriendly>=1.44.3
influxdb>=2.12.0
numpy>=1.11.1
pyyaml>=3.11
